+++ src/main/java/net/sf/jabref/gui/maintable/MainTableNameFormatter.java
@@ -13,6 +13,10 @@ public class MainTableNameFormatter {
      * @return The formatted name field.
      */
     public static String formatName(String nameToFormat) {
+        if (nameToFormat == null) {
+            return null;
+        }
+
         // Read name format options:
         boolean namesNatbib = Globals.prefs.getBoolean(JabRefPreferences.NAMES_NATBIB); //MK:
         boolean namesLastOnly = Globals.prefs.getBoolean(JabRefPreferences.NAMES_LAST_ONLY);
@@ -21,9 +25,7 @@ public class MainTableNameFormatter {
         boolean namesFf = Globals.prefs.getBoolean(JabRefPreferences.NAMES_FIRST_LAST);
         boolean namesLf = !(namesAsIs || namesFf || namesNatbib || namesLastOnly); // None of the above.
 
-        if (nameToFormat == null) {
-            return null;
-        } else if (namesAsIs) {
+        if (namesAsIs) {
             return nameToFormat;
         } else if (namesNatbib) {
             nameToFormat = AuthorList.fixAuthor_Natbib(nameToFormat);